requirements.txt
iniconfig==2.0.0
packaging==23.2
pluggy==1.4.0
pytest==8.0.0
python-environ==0.4.54
